% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{config_df}
\alias{config_df}
\title{Default parameters of config.}
\format{A data frame with 12 variables:
\describe{
  \item{\code{threshold}}{Threshold for allele frequency}
  \item{\code{skew}}{Skewness for allele frequency}
  \item{\code{lower}}{Lower bound for allele frequency region}
  \item{\code{upper}}{Upper bound for allele frequency region}
  \item{\code{ldpthred}}{Threshold to determine low depth}
  \item{\code{hom_mle}}{Hom MLE of p in Beta-Binomial model}
  \item{\code{het_mle}}{Het MLE of p in Beta-Binomial model}
  \item{\code{Hom_thred}}{Threshold between hom and high}
  \item{\code{High_thred}}{Threshold between high and het}
  \item{\code{Het_thred}}{Threshold between het and low}
  \item{\code{hom_rho}}{Hom MLE of rho in Beta-Binomial model}
  \item{\code{het_rho}}{Het MLE of rho in Beta-Binomial model}
}}
\source{
Created by Tao Jiang
}
\usage{
config_df
}
\description{
A dataframe containing default parameters.
}
\keyword{datasets}
