% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_feature.R
\name{getAnnoRate}
\alias{getAnnoRate}
\title{Annotation rate}
\usage{
getAnnoRate(f)
}
\arguments{
\item{f}{Input raw file}
}
\value{
Percentage of annotation locus
}
\description{
Annotation rate
}
