% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_feature.R
\name{getAvgLL}
\alias{getAvgLL}
\title{Calculate average log-likelihood}
\usage{
getAvgLL(df, hom_mle, het_mle, hom_rho, het_rho)
}
\arguments{
\item{df}{Input modified file}

\item{hom_mle}{Hom MLE of p in Beta-Binomial model, default is 0.9981416 from NA12878_1_L5}

\item{het_mle}{Het MLE of p in Beta-Binomial model, default is 0.4737897 from NA12878_1_L5}

\item{hom_rho}{Hom MLE of rho in Beta-Binomial model, default is 0.04570275 from NA12878_1_L5}

\item{het_rho}{Het MLE of rho in Beta-Binomial model, default is 0.02224098 from NA12878_1_L5}
}
\value{
meanLL
}
\description{
Calculate average log-likelihood
}
