% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_feature.R
\name{getLowDepth}
\alias{getLowDepth}
\title{Low depth percentage}
\usage{
getLowDepth(f, ldpthred)
}
\arguments{
\item{f}{Input raw file}

\item{ldpthred}{Threshold to determine low depth, default is 20}
}
\value{
Percentage of low depth
}
\description{
Low depth percentage
}
