% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_vcf.R
\name{getRatio}
\alias{getRatio}
\title{Get the ratio of allele frequencies with a region}
\usage{
getRatio(subdf, lower, upper)
}
\arguments{
\item{subdf}{Dataframe with calculated statistics}

\item{lower}{Lower bound for allele frequency region}

\item{upper}{Upper bound for allele frequency region}
}
\value{
Ratio of allele frequencies with a region
}
\description{
Get the ratio of allele frequencies with a region
}
