% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_vcf.R
\name{getSkewness}
\alias{getSkewness}
\title{Get absolute value of skewness}
\usage{
getSkewness(subdf)
}
\arguments{
\item{subdf}{Input dataframe}
}
\value{
Absolute value of skewness
}
\description{
Get absolute value of skewness
}
