% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{locateFile}
\alias{locateFile}
\title{Check input filename}
\usage{
locateFile(fn, extension)
}
\arguments{
\item{fn}{Exact full file name of input file, including directory}

\item{extension}{Expected input file extension: vcf & txt}
}
\value{
Valid directory
}
\description{
Check input filename
}
