% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{readVarPROWL}
\alias{readVarPROWL}
\title{Read in input vcf data in VarPROWL format}
\usage{
readVarPROWL(dr, dbOnly, depCut, thred, content, extnum, keepall)
}
\arguments{
\item{dr}{A valid input object}

\item{dbOnly}{Use dbSNP as filter, default is FALSE, passed from read_vcf}

\item{depCut}{Use a threshold for min depth , default is False}

\item{thred}{Threshold for min depth, default is 20}

\item{content}{Column names in VCF files}

\item{extnum}{The column number or numbers to be extracted from vcf, default is 10; 0 for not extracting any columns}

\item{keepall}{Keep unextracted column in output, default is TRUE, passed from read_vcf}
}
\value{
vcf Dataframe from VCF file
}
\description{
Read in input vcf data in VarPROWL format
}
