% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_vcf.R
\name{rmChangePoint}
\alias{rmChangePoint}
\title{Remove CNV regions within VCF files by changepoint method}
\usage{
rmChangePoint(vcf, threshold, skew, lower, upper)
}
\arguments{
\item{vcf}{Input VCF files}

\item{threshold}{Threshold for allele frequency}

\item{skew}{Skewness for allele frequency}

\item{lower}{Lower bound for allele frequency region}

\item{upper}{Upper bound for allele frequency region}
}
\value{
VCF object without changepoint region
}
\description{
Remove CNV regions within VCF files by changepoint method
}
