% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_vcf.R
\name{rmCNVinVCF}
\alias{rmCNVinVCF}
\title{Remove CNV regions within VCF files given cnv file}
\usage{
rmCNVinVCF(vcf, cnvobj)
}
\arguments{
\item{vcf}{Input VCF files}

\item{cnvobj}{cnv object}
}
\value{
VCF object without changepoint region
}
\description{
Remove CNV regions within VCF files given cnv file
}
