% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{svm_class_model}
\alias{svm_class_model}
\title{Default svm classification model.}
\format{An svm object:}
\source{
Created by Tao Jiang
}
\usage{
svm_class_model
}
\description{
An svm object containing default svm classification model.
}
\keyword{datasets}
