% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{svm_regression_model}
\alias{svm_regression_model}
\title{Default svm regression model.}
\format{An svm object:}
\source{
Created by Tao Jiang
}
\usage{
svm_regression_model
}
\description{
An svm object containing default svm regression model.
}
\keyword{datasets}
